requirements.txt
Babel==2.13.1
beautifulsoup4==4.12.2
bracex==2.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
exceptiongroup==1.2.0
future==0.18.3
geocoder==1.38.1
ghp-import==2.1.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
libsass==0.22.0
livereload==2.6.3
Markdown==3.3.7
markdown-customblocks==1.4.1
MarkupSafe==2.1.3
material==0.1
mergedeep==1.3.4
mkdocs==1.6
mkdocs-extra-sass-plugin==0.1.0
mkdocs-glightbox==0.3.5
mkdocs-include-markdown-plugin==6.0.3
mkdocs-material==9.5.34
mkdocs-material-extensions==1.3
mkdocs-toc-md==0.0.8
packaging==23.1
paginate==0.5.6
pathspec==0.11.2
platformdirs==4.1.0
pluggy==1.3.0
Pygments==2.16.0
pymdown-extensions==10.2
pytest==7.4.3
python-dateutil==2.8.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
ratelim==0.1.6
regex==2023.6.3
requests==2.31.0
responses==0.24.1
six==1.16.0
soupsieve==2.4.1
tomli==2.0.1
tornado==6.3.2
urllib3==1.26.18
watchdog==3.0.0
wcmatch==8.5
yamlns==0.11.0
